Make driver_needs_utf8 configurable in OPTIONS.
Recent versions of pyodbc and FreeTDS (3.0.6 and 0.91 as of writing) are
perfectly okay with being passed unicode. While I'm not entirely sure of the
specifics, in testing everything is okay and there's no need to encode to utf8
in this scenario, and so now the option is configurable. The default is still
to encode on drivers other than Windows drivers, but at least it's configurable.