Encode connection strings to escape semicolon etc
ODBC connection string values that contain a semicolon must be escaped, as the
semicolon is the separator used between KeyValuePairs. So build up the
connection string in a structured way, using a dict to store KeyValuePairs so
we can be sure to properly encode all the values used, even those that come
from Django settings where the raw, unescaped values will be given.